+++ stories/templates/switch/default.template.js
@@ -1,0 +1,6 @@
+export const template = `
+    <st-switch :inactive-label="inactiveText"
+               :active-label="activeText"
+               :disabled="disabled"
+               v-model="value"></st-switch>
+`;